add int entities

+++ src/productos/entities/producto.entity.ts
@@ -17,7 +17,7 @@ export class Producto {
   @Column()
   nombre: string;
 
-  @Column({ type: "double", precision: 10, scale: 2 })
+  @Column({ type: "int", precision: 10, scale: 2 })
   precio: number;
 
   @Column()
